Extract Superclass, Encapsulate Field
1.BlacklistFolderChooserDialog, ScanMediaFolderChooserDialog, ChooserDialog
2. Extract Superclass, Encapsulate Field
3. 먼저 두 클래스 간의 공통되는 메서드가 많아 Extract superclass를 해주었습니다. 그리고 superclass의 필드는 encapsulation을 위해서 private으로 두고 getter/setter 메서드를 만들어 적용했습니다.
