+++ app/src/main/java/com/kabouzeid/gramophone/dialogs/BlacklistFolderChooserDialog.java
@@ -23,52 +23,10 @@ import java.util.List;
 /**
  * @author Aidan Follestad (afollestad), modified by Karim Abou Zeid
  */
-public class BlacklistFolderChooserDialog extends DialogFragment implements MaterialDialog.ListCallback {
-
-    private File parentFolder;
-    private File[] parentContents;
-    private boolean canGoUp = false;
+public class BlacklistFolderChooserDialog extends ChooserDialog implements MaterialDialog.ListCallback {
 
     private FolderCallback callback;
 
-    String initialPath = Environment.getExternalStorageDirectory().getAbsolutePath();
-
-    private final boolean isSDKAboveAndroidMarshmallow = Build.VERSION.SDK_INT >= Build.VERSION_CODES.M;
-    private final boolean isNotGrantedPermissionToReadExternalStorage = ActivityCompat.checkSelfPermission(
-            getActivity(), Manifest.permission.READ_EXTERNAL_STORAGE)
-            != PackageManager.PERMISSION_GRANTED;
-
-    private String[] getContentsArray() {
-        if (parentContents == null) {
-            if (canGoUp) {
-                return new String[]{".."};
-            }
-            return new String[]{};
-        }
-        String[] results = new String[parentContents.length + (canGoUp ? 1 : 0)];
-        if (canGoUp) {
-            results[0] = "..";
-        }
-        for (int i = 0; i < parentContents.length; i++) {
-            results[canGoUp ? i + 1 : i] = parentContents[i].getName();
-        }
-        return results;
-    }
-
-    private File[] listFiles() {
-        File[] contents = parentFolder.listFiles();
-        List<File> results = new ArrayList<>();
-        if (contents != null) {
-            for (File fi : contents) {
-                if (fi.isDirectory()) {
-                    results.add(fi);
-                }
-            }
-            Collections.sort(results, new FolderSorter());
-            return results.toArray(new File[results.size()]);
-        }
-        return null;
-    }
 
     public static BlacklistFolderChooserDialog create() {
         return new BlacklistFolderChooserDialog();
@@ -91,18 +49,20 @@ public class BlacklistFolderChooserDialog extends DialogFragment implements Mate
         if (!savedInstanceState.containsKey("current_path")) {
             savedInstanceState.putString("current_path", initialPath);
         }
-        parentFolder = new File(savedInstanceState.getString("current_path", File.pathSeparator));
+
+        setParentFolder(new File(savedInstanceState.getString("current_path", File.pathSeparator)));
+
         checkIfCanGoUp();
-        parentContents = listFiles();
+        setParentContents(listFiles());
         MaterialDialog.Builder builder =
                 new MaterialDialog.Builder(getActivity())
-                        .title(parentFolder.getAbsolutePath())
+                        .title(getParentFolder().getAbsolutePath())
                         .items((CharSequence[]) getContentsArray())
                         .itemsCallback(this)
                         .autoDismiss(false)
                         .onPositive((dialog, which) -> {
                             dismiss();
-                            callback.onFolderSelection(BlacklistFolderChooserDialog.this, parentFolder);
+                            callback.onFolderSelection(BlacklistFolderChooserDialog.this, getParentFolder());
                         })
                         .onNegative((materialDialog, dialogAction) -> dismiss())
                         .positiveText(R.string.add_action)
@@ -113,37 +73,21 @@ public class BlacklistFolderChooserDialog extends DialogFragment implements Mate
     @Override
     public void onSelection(MaterialDialog materialDialog, View view, int i, CharSequence s) {
         if (canGoUp && i == 0) {
-            parentFolder = parentFolder.getParentFile();
-            if (parentFolder.getAbsolutePath().equals("/storage/emulated")) {
-                parentFolder = parentFolder.getParentFile();
+            setParentFolder(getParentFolder().getParentFile());
+            if (getParentFolder().getAbsolutePath().equals("/storage/emulated")) {
+                setParentFolder(getParentFolder().getParentFile());
             }
             checkIfCanGoUp();
         } else {
-            parentFolder = parentContents[canGoUp ? i - 1 : i];
+            setParentFolder(getParentContents()[canGoUp ? i - 1 : i]);
             canGoUp = true;
-            if (parentFolder.getAbsolutePath().equals("/storage/emulated")) {
-                parentFolder = Environment.getExternalStorageDirectory();
+            if (getParentFolder().getAbsolutePath().equals("/storage/emulated")) {
+                setParentFolder(Environment.getExternalStorageDirectory());
             }
         }
         reload();
     }
 
-    private void checkIfCanGoUp() {
-        canGoUp = parentFolder.getParent() != null;
-    }
-
-    private void reload() {
-        parentContents = listFiles();
-        MaterialDialog dialog = (MaterialDialog) getDialog();
-        dialog.setTitle(parentFolder.getAbsolutePath());
-        dialog.setItems((CharSequence[]) getContentsArray());
-    }
-
-    @Override
-    public void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-        outState.putString("current_path", parentFolder.getAbsolutePath());
-    }
 
     public void setCallback(FolderCallback callback) {
         this.callback = callback;
@@ -153,11 +97,5 @@ public class BlacklistFolderChooserDialog extends DialogFragment implements Mate
         void onFolderSelection(@NonNull BlacklistFolderChooserDialog dialog, @NonNull File folder);
     }
 
-    private static class FolderSorter implements Comparator<File> {
 
-        @Override
-        public int compare(File lhs, File rhs) {
-            return lhs.getName().compareTo(rhs.getName());
-        }
-    }
 }
